requirements.txt
absl-py
astor
atomicwrites
attrs
cloudpickle
colorama
coverage
cycler
dask
decorator
docopt
gast
grpcio
h5py
Keras
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
more-itertools
networkx
numpy
packaging
pandas
Pillow
pluggy
protobuf
py
pyparsing
pytest
python-dateutil
pytz
PyWavelets
PyYAML
scikit-image
scikit-learn
scipy
six
sklearn
tensorboard
tensorflow
termcolor
toolz
tox
tqdm
virtualenv
Werkzeug